refactor: move server implementation to internal/server package
