+++ internal/server/cors.go
@@ -1,4 +1,4 @@
-package nine
+package server
 
 import (
 	"fmt"
