+++ internal/server/handler_validator_test.go
@@ -1,4 +1,4 @@
-package nine
+package server
 
 import (
 	"errors"
