+++ internal/server/route_group.go
@@ -1,4 +1,4 @@
-package nine
+package server
 
 import (
 	"strings"
