+++ internal/server/route_group_test.go
@@ -1,4 +1,4 @@
-package nine
+package server
 
 import (
 	"net/http"
@@ -6,15 +6,17 @@ import (
 	"testing"
 
 	"github.com/i9si-sistemas/assert"
+	"github.com/i9si-sistemas/nine/internal/json"
 )
 
 func TestRouteGroup(t *testing.T) {
-	testServer := NewServer(8080)
+	testServer := New(8080)
 	type Account struct {
 		Name string `json:"name"`
 		Age  int    `json:"age"`
 	}
-	accounts := make(GenericJSON[string, Account], 0)
+	type JSON map[string]any
+	accounts := make(map[string]Account, 0)
 	testServer.Route("/account", func(router *RouteGroup) {
 		router.Post("/create", func(c *Context) error {
 			var body Account
@@ -49,30 +51,30 @@ func TestRouteGroup(t *testing.T) {
 	var account struct {
 		Account `json:"account"`
 	}
-	err := DecodeJSON(b, &account)
+	err := json.Decode(b, &account)
 	assert.NoError(t, err)
 	assert.Equal(t, account.Name, "Gabriel Luiz")
 	assert.Equal(t, account.Age, 23)
 }
 
 func TestGroup(t *testing.T) {
-	testServer := NewServer(5024)
+	testServer := New(5024)
 	type Account struct {
 		Name  string `json:"name"`
 		Age   int    `json:"age"`
 		Money int64  `json:"money"`
 	}
-	accounts := make(GenericJSON[string, Account], 0)
+	accounts := make(JSON[Account], 0)
 	accountGroup := testServer.Group("/account")
 	accountGroup.Post("/create", func(c *Context) error {
 		var body Account
 		if err := Body(c.Request, &body); err != nil {
-			return c.Status(http.StatusBadRequest).JSON(JSON{
+			return c.Status(http.StatusBadRequest).JSON(JSON[string]{
 				"message": "invalid body",
 			})
 		}
 		_, ok := accounts[body.Name]
-		response := JSON{"created": !ok}
+		response := JSON[bool]{"created": !ok}
 		if !ok {
 			accounts[body.Name] = body
 			return c.Status(http.StatusCreated).JSON(response)
@@ -85,7 +87,7 @@ func TestGroup(t *testing.T) {
 		if !ok {
 			return c.SendStatus(http.StatusNotFound)
 		}
-		return c.JSON(JSON{
+		return c.JSON(JSON[Account]{
 			"account": acc,
 		})
 	})
@@ -102,7 +104,7 @@ func TestGroup(t *testing.T) {
 	var account struct {
 		Account `json:"account"`
 	}
-	err := DecodeJSON(b, &account)
+	err := json.Decode(b, &account)
 	assert.NoError(t, err)
 	assert.Equal(t, account.Name, "Gabriel Luiz")
 	assert.Equal(t, account.Age, 23)
@@ -118,7 +120,7 @@ func assertGroupEndpoints(t assert.T, testServer *Server) {
 	var response struct {
 		Created bool `json:"created"`
 	}
-	payload, err := JSON{
+	payload, err := JSON[any]{
 		"name":  "Gabriel Luiz",
 		"age":   23,
 		"money": 5000,
@@ -128,11 +130,11 @@ func assertGroupEndpoints(t assert.T, testServer *Server) {
 	w := testServer.Test().Request(req)
 	assert.Equal(t, w.Result().StatusCode, http.StatusCreated)
 	b := w.Body.Bytes()
-	err = DecodeJSON(b, &response)
+	err = json.Decode(b, &response)
 	assert.NoError(t, err)
 	assert.True(t, response.Created)
 
-	payload, err = JSON{
+	payload, err = JSON[any]{
 		"name":  "Gabriel Luiz",
 		"age":   23,
 		"money": 5000000,
@@ -142,7 +144,7 @@ func assertGroupEndpoints(t assert.T, testServer *Server) {
 	w = testServer.Test().Request(req)
 	assert.Equal(t, w.Result().StatusCode, http.StatusOK)
 	b = w.Body.Bytes()
-	err = DecodeJSON(b, &response)
+	err = json.Decode(b, &response)
 	assert.NoError(t, err)
 	assert.False(t, response.Created)
 }
